+++ CHANGELOG.rst
@@ -2,8 +2,16 @@
 Changelog
 =========
 
+0.x.x (?)
+==================
+
+* Added ...
+* Added Minimum option for Timeseries
+* Added Maximum option for Timeseries
+* Added Number of decimals displays option for Timeseries
+
 0.7.0 (2022-10-02)
-===========
+==================
 
 * Added Grafana 8.x new Alert Rule
 * Added Grafana 9.x new Alert Rule
@@ -16,7 +24,6 @@ Changelog
 * Fix typo in unit constant ``NORMAL_CUBIC_METER`` (was ``NORMAIL_CUBIC_METER``)
 
 .. _ePict: basehttps://grafana.com/grafana/plugins/larona-epict-panel/
-
 
 0.6.3 (2022-03-30)
 ==================